requirements.txt
alabaster==0.7.12
alembic==1.0.7
asn1crypto==0.24.0
astroid==2.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
cryptography==2.5
decorator==4.3.2
docutils==0.14
dominate==2.3.5
entrypoints==0.3
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
jeepney==0.4
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==17.1.1
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1.1
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.15.4
numpydoc==0.8.0
olefile==0.46
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.2
pdfminer.six==20170720
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
psutil==5.5.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.1.0
Pygments==2.3.1
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2018.9
pyzmq==17.1.2
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
requests==2.21.0
rope==0.11.0
SecretStorage==3.1.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.2.18
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
wurlitzer==1.0.2
